app/requirements.txt
fastapi
uvicorn
python-dotenv
prometheus_client
psutil